requirements.txt
# Core Libraries
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
pillow>=10.0.0

# GUI
PyQt5>=5.15.0

# Deep Learning (CUDA support)
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0

# VLM Integration
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0

# Utilities
tqdm>=4.66.0
